+++ examples/resources/sonatyperepo_security_ssrf_protection/resource.tf
@@ -1,0 +1,9 @@
+resource "sonatyperepo_security_ssrf_protection" "ssrf" {
+  enabled = true
+  allowed_domains = [
+    "internal.example.com"
+  ]
+  allowed_ips = [
+    "10.0.0.5"
+  ]
+}